Move MergedPullRequestFinder into own file
